Refactoring base_page. Add home_pageed
