+++ dp189/components.py
@@ -1,0 +1,108 @@
+from selenium.webdriver.common.keys import Keys
+from selenium.webdriver import Remote
+from .locators import LocatorsSearch, LocatorsNavBar, RightMenuLocators, LocatorsShoppingCartButton
+
+class SearchArea:
+    """"This class describes the search area common in all pages. It consists or search field and search button"""
+    def __init__(self, driver):
+        self._driver = driver
+        self._search_field = driver.find_element(*LocatorsSearch.SEARCH_FIELD)
+        self._search_button = driver.find_element(*LocatorsSearch.SEARCH_BUTTON)
+
+    
+    def fill_search_field_and_hit_return(self, item: str):
+        self._search_field.clear()
+        self._search_field.send_keys(item).send_keys(Keys.RETURN)
+        #return SearchPage(self._driver)
+    
+    def fill_search_field_and_click(self, item: str):
+        self._search_field.clear()
+        self._search_field.send_keys(item)
+        self._search_button.click()
+        #return SearchPage(self._driver)
+
+
+class ShopCartButton:
+    # TODO test functionality
+    def __init__(self, driver):
+        self._driver = driver
+        self._shop_cart_button = driver.find_element(*LocatorsShoppingCartButton.SHOP_CART_BUTTON)
+
+    def click_shop_cart_button(self):
+        self._shop_cart_button.click()
+        cart_items = self._driver.find_elements(*LocatorsShoppingCartButton.CART_ITEMS)
+        if len(cart_items) == 0:
+            return 'Your cart is empty!'
+        else:
+            return ShopCartDropdown(self._driver)
+
+
+class ShopCartDropdown:
+    def __init__(self, driver):
+        pass
+
+
+class BaseNavBar:
+    """This class describes the top nav bar of the base page"""
+    def __init__(self, driver):
+        self._driver = driver
+        self._currency = driver.find_element(*LocatorsNavBar.CURRENCY)
+        self._nav_bar = driver.find_element(*LocatorsNavBar.NAVBAR)
+
+    def click_currency_euro(self):
+        self._currency.click()
+        self._currency.find_element(*LocatorsNavBar.EUR).click()
+
+    def click_currency_pound(self):
+        self._currency.click()
+        self._currency.find_element(*LocatorsNavBar.POUND).click()
+
+    def click_currency_usd(self):
+        self._currency.click()
+        self._currency.find_element(*LocatorsNavBar.USD).click()
+
+    def click_contact_us(self):
+        self._nav_bar.find_element(*LocatorsNavBar.CONTACT_US).click()
+
+    def click_wishlist(self):
+        self._nav_bar.find_element(*LocatorsNavBar.WISH_LIST).click()
+
+    def click_shopping_cart(self):
+        self._nav_bar.find_element(*LocatorsNavBar.SHOPPING_CART).click()
+
+
+class BaseRightMenu:
+    def __init__(self, driver) -> None:
+        self._driver = driver
+        self._right_menu = driver.find_element_by_class_name('list-group')
+
+    def click_my_account(self):
+        self._right_menu.find_element(*RightMenuLocators.MY_ACCOUNT).click()
+
+    def click_address_book(self):
+        self._right_menu.find_element(*RightMenuLocators.ADDRESS_BOOK).click()
+
+    def click_wish_list(self):
+        self._right_menu.find_element(*RightMenuLocators.WISH_LIST).click()
+
+    def click_order_history(self):
+        self._right_menu.find_element(*RightMenuLocators.ORDER_HISTORY).click()
+
+    def click_downloads(self):
+        self._right_menu.find_element(*RightMenuLocators.DOWNLOADS).click()
+
+    def click_recurring_payments(self):
+        self._right_menu.find_element(*RightMenuLocators.RECURRING_PAYMENTS).click()
+
+    def click_reward_points(self):
+        self._right_menu.find_element(*RightMenuLocators.REWARD_POINTS).click()
+
+    def click_returns(self):
+        self._right_menu.find_element(*RightMenuLocators.RETURNS).click()
+
+    def click_transactions(self):
+        self._right_menu.find_element(*RightMenuLocators.TRANSACTIONS).click()
+
+    def click_newsletter(self):
+        self._right_menu.find_element(*RightMenuLocators.NEWSLETTER).click()
+
